+++ Shared.Abstract/Primitives/AggregateRoot.cs
@@ -1,0 +1,13 @@
+﻿namespace TaskNotFound.Shared.Abstract.Primitives;
+
+public abstract class AggregateRoot<TId>(TId id) : Entity<TId>(id)
+    where TId : notnull
+{
+    private readonly List<IDomainEvent> _domainEvents = [];
+
+    public IReadOnlyCollection<IDomainEvent> DomainEvents => _domainEvents.AsReadOnly();
+
+    public void RaiseDomainEvent(IDomainEvent domainEvent) => _domainEvents.Add(domainEvent);
+
+    public void ClearDomainEvents() => _domainEvents.Clear();
+}