+++ Shared.Abstract/Primitives/Entity.cs
@@ -1,0 +1,33 @@
+﻿namespace TaskNotFound.Shared.Abstract.Primitives;
+
+public abstract class Entity<TId>(TId id) : IEquatable<Entity<TId>>
+    where TId : notnull
+{
+    public TId Id { get; } = id;
+
+    public bool Equals(Entity<TId>? other)
+    {
+        if (this == other)
+        {
+            return true;
+        }
+
+        return other is not null && Id.Equals(other.Id);
+    }
+
+    public override bool Equals(object? obj)
+    {
+        if (obj is null)
+        {
+            return false;
+        }
+
+        return obj is Entity<TId> entity && Equals(entity);
+    }
+
+    public override int GetHashCode() => Id.GetHashCode();
+
+    public static bool operator ==(Entity<TId>? left, Entity<TId>? right) => Equals(left, right);
+
+    public static bool operator !=(Entity<TId>? left, Entity<TId>? right) => !Equals(left, right);
+}