+++ Shared.Abstract/Primitives/IDomainEvent.cs
@@ -1,0 +1,3 @@
+﻿namespace TaskNotFound.Shared.Abstract.Primitives;
+
+public interface IDomainEvent;